Changing RGB to BGRA pixel format in imageIO
jpg, png, tiff done

+++ src/utility/imageio/png.h
@@ -1,5 +1,5 @@
 /*  Created,  18 Jul 2018   */
-/*  Last Edit 20 Jul 2018   */
+/*  Last Edit  4 Aug 2018   */
 
 /**
  *  \file png.h
@@ -8,7 +8,7 @@
  *  aid of libpng
  *  \author Davide Pizzolotto
  *  \version 0.2
- *  \date 20 Jul 2018
+ *  \date  4 Aug 2018
  *  \copyright MIT
  */
 
@@ -32,11 +32,12 @@
  *  Existing files will be overwritten
  *  \param[in] width Width of the image
  *  \param[in] height Height of the image
- *  \param[in] data array of 8-bit values containing the image data in
- *  top-down left to right order. The values are in order RGBRGB...
+ *  \param[in] data array of 32-bit values containing the image data in
+ *  top-down left to right order. The channels for each pixel are in BGRA order,
+ *  with the alpha value always 255.
  *  \return 1 if the image was successfully written, 0 otherwise
  */
-char png_write(const char* name, int width, int height, const uint8_t* data);
+char png_write(const char* name, int width, int height, const uint32_t* data);
 
 /**
  *  \brief Checks if a png image is valid
@@ -71,8 +72,8 @@ char png_dimensions(const char* name, int* width, int* height);
  *
  *  Reads a png image and writes the image data on the values array in
  *  top-down left-right order. This array is expected to be already allocated
- *  with a size of width*height*3. Every channel is written as a single uint8_t
- *  value in the order RGBRGB... with values in the range [0-255].
+ *  with a size of width*height. Every channel is written in a single uint32_t
+ *  value BGRA with values in the range [0-255] and alpha value always 255.
  *
  *  \param[in] name The path on disk where the image can be found
  *  \param[out] values The array of values that will be written. This array
@@ -80,6 +81,6 @@ char png_dimensions(const char* name, int* width, int* height);
  *  the image and to preallocate it
  *  \return 0 if the read was not successful, 1 otherwise
  */
-char png_read(const char* name, uint8_t* values);
+char png_read(const char* name, uint32_t* values);
 
 #endif
